requirements.txt
fitz
google-genai
sentence-transformers
langchain
chromadb
langchain-text-splitters
